mail-auth-test/requirements.txt
python-dotenv

# formatter
black
isort